requirements.txt
PyQt6>=6.6
pyserial>=3.5
